% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{add2labCatSelectionDBtable}
\alias{add2labCatSelectionDBtable}
\title{add2labCatSelectionDBtable}
\usage{
add2labCatSelectionDBtable(
  df.ref = "reference table",
  cat_group_name = "Tybulewicz lab",
  reference.gene.vector = "reference gene vector",
  ref.gene.vec.id = "hgnc_symbol",
  cat_views = NA
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
