% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{add2vec}
\alias{add2vec}
\title{A method}
\usage{
add2vec(obj, slot_name, value)
}
\arguments{
\item{obj}{A bioLOGIC data object}

\item{slot_name}{character vector slot (e.g. reportVec, scriptVec or documentationVector) to which an element is to be added}

\item{value}{Element or string to be added to the above vector in the bioLOGIC object}
}
\description{
Method description
}
