% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{addCoVar}
\alias{addCoVar}
\title{A method}
\usage{
addCoVar(
  obj,
  avgCountCutOffperSample = 0,
  selectionColName = "aboveCutOff",
  dfBaseData = "Obio@DESeqNormReadCountsTable",
  rowNameID = "obj@parameterList$primaryAlignmentGeneID"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
