% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{addDGEcomparisons2DesignFile}
\alias{addDGEcomparisons2DesignFile}
\title{A method}
\usage{
addDGEcomparisons2DesignFile(
  dfDesign = "dfDesign",
  comparisonList = "list(\\n    \\"WT_vs_MT\\" = c(\\"MT_YapTaz\\", \\"WT_WT\\")\\n)"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
