% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{addGeneAnnotation}
\alias{addGeneAnnotation}
\title{A method}
\usage{
addGeneAnnotation(obj, addUniprotColumn = FALSE)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
