% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{addProject2ProjectTable}
\alias{addProject2ProjectTable}
\title{addProject2ProjectTable}
\usage{
addProject2ProjectTable(
  dbname = "reference_categories_db_new",
  dbtable = "project_db_table",
  password = db.pwd,
  experiment_id = project_id,
  experiment_question = "Which genes are candidates for G2M checkpoint regulation?",
 
    experiment_description = "<b>RNAi screen</b> for genes that affect the G2M checkpoint",
  experiment_owner = ";Katharina Deiss;Nicola Longwood;",
  experiment_lab = paste0(";", labname, ";"),
  experiment_viewers = paste0(";", paste(user_ids, collapse = ";"), ";"),
  experiment_project = ";G2 Cell Cycle Checkpoint;",
  experiment_type = experiment.type,
  experiment_details = "Experimental details will follow here",
  experiment_x_coordinate = ";Screen;",
  experiment_x_coordinate_unit = "condition",
  experiment_link = paste0("<a href='../", project_id,
    "/report.php' class='btn btn-success btn-lg' role='button'>Primary Data &raquo;</a>"),
  experiment_title = ""
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
