% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{addTPMorFPKMtable}
\alias{addTPMorFPKMtable}
\title{addTPMorFPKMtable}
\usage{
addTPMorFPKMtable(obj = "biologic objec", addTPM = TRUE, addFPKM = FALSE)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
