% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\docType{class}
\name{bioLOGIC-class}
\alias{bioLOGIC-class}
\title{bioLOGIC}
\arguments{
\item{agree}{Do you agree dogs are the best pet? Defaults to TRUE.}
}
\description{
This function allows you to express your love for the superior furry animal.
}
\keyword{dogs}
