% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{completeDesignBasedOnSampleID}
\alias{completeDesignBasedOnSampleID}
\title{A method}
\usage{
completeDesignBasedOnSampleID(dfBasedesign, fastqDir = "")
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
