% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{create.gmt.file.from.ref.data.table}
\alias{create.gmt.file.from.ref.data.table}
\title{A method}
\usage{
create.gmt.file.from.ref.data.table(
  host = "www.biologic-db.org",
  port = port,
  dbname = "reference_categories_db_new",
  dataTable = "st_lab_categories",
  pwd = "Saturday08",
  user = "boeings",
  gene.id.column = "hgnc_symbol"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
