% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{create.gsea.rnk.files}
\alias{create.gsea.rnk.files}
\title{A method}
\usage{
create.gsea.rnk.files(
  localWorkDir,
  df.dataTable,
  GSEA.colum.type = "stat",
  gene.symbol.column.name = "hgnc_symbol",
  medianCollapse = TRUE,
  GSEADir = NULL
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
