% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{create.GSEA.table}
\alias{create.GSEA.table}
\title{A method}
\usage{
create.GSEA.table(
  GSEADir,
  logFC.column.name = "logFC",
  host = "www.biologic-db.org",
  refdbname = "reference_categories_db_new",
  refDBTableName = paste0(project.code, "_VTL_ES_enriched_categories_table"),
  db.user = db.user,
  db.password = db.pwd,
  tables = tables,
  df.dataTable,
  outputDir = "",
  project.code = "p001"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
