% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{create.rnaseqc.script}
\alias{create.rnaseqc.script}
\title{A method}
\usage{
create.rnaseqc.script(
  df.design,
  sample.column = "sample.id",
  project.code = "p103",
  project = "SB_RNAseqQC",
 
    basedataDir = "/camp/stp/babs/working/boeings/Projects/103_VTL_ES_RNA_seq_BAFF_timecourse_hs/workdir/RSEM/Ensembl",
  bam.suffix = "STAR.genome.bam",
 
    GTFfile = "/camp/stp/babs/working/data/genomes/homo_sapiens/ensembl/GRCh38/release-86/gtf/Homo_sapiens.GRCh38.86.rnaseqc.gtf",
 
    rRNAfile = "/camp/stp/babs/working/data/genomes/homo_sapiens/ensembl/GRCh38/release-86/gtf/Homo_sapiens.GRCh38.86.rRNA.list",
 
    genome.fa = "/camp/stp/babs/working/data/genomes/homo_sapiens/ensembl/GRCh38/release-86/genome/Homo_sapiens.GRCh38.dna_sm.primary_assembly.fa",
  refFlatFile = "",
  ribosomalIntervalList = "",
  bedFile = "",
  paired.end = FALSE,
  strandSpecific = TRUE
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
\keyword{n=n-1',}
