% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{create.timecourse.cat.lines}
\alias{create.timecourse.cat.lines}
\title{create.timecourse.cat.lines}
\usage{
create.timecourse.cat.lines(
  df.design,
  display.median.string = "calculate_median"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
