% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{create.tpm.and.fpkm.tables}
\alias{create.tpm.and.fpkm.tables}
\title{A method}
\usage{
create.tpm.and.fpkm.tables(workdir, samples, files)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
