% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{create.website.parameters}
\alias{create.website.parameters}
\title{create.website.parameters}
\usage{
create.website.parameters(
  df.data,
  gene.id.column = "hgnc_symbol",
  ptm.colum = "display_ptm",
  lab_id = "st_lab",
  user_ids = c("project", "st_lab_all", "thomas.mercer"),
  project_id = "stl1",
  download_result_table = "20160823.logFC.datatable.txt",
  download_cat_enrichment_table = "p50.interesting.categories.txt",
  database = "stl_data",
  reference_categories_db = "reference_categories_db_new",
  labname = "Tooze",
  rnaseqdbTableName,
  lab.categories.table = "st_lab_categories",
  sample.order = names(df.data)[grepl("norm_counts_", names(df.data))],
  count.sample.colors = "",
  count.table.headline = "PTM ratio H/L counts for all samples",
  count.column.chart.x.axis.label = "hrs",
  count.table.sidelabel = "Counts",
  webSiteDir = "C:/xampp/htdocs/",
  heamap.headline.text = "log2FC(SILAC) Heatmap",
  upper_heatmap_limit = 3,
  lower_heatmap_limit = -3,
  slider.selection.name = "logFC",
  presentation.file = "",
  number_of_slides = "",
  default.sequence = "_____ARK_______",
  use.logFC.columns.for.heatmap = FALSE,
  peptide.view.link = "",
  create.2d.scatterplot.button = FALSE,
  low_highlight = -1,
  high_highlight = 1,
  display.qc = FALSE,
  display.pca = FALSE,
  display.report = FALSE,
  pca.table.name = "",
  gsea.cat.lines = NA,
  timecourse.cat.lines = NA,
  venn.slider.selector.strings = c("contrast_x_logFC", "constrast_x_padj"),
  plot.selection.strings = NA,
  plate.view.db.table = NA,
  plate.view.column.vec = NA,
  cat.seletion.table.vec = c("reference_categories_db_new", "cat_selection_default", ),
  localhost = "localhost",
  createOutputFile = TRUE
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
