% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createAlignmentShellScript}
\alias{createAlignmentShellScript}
\title{A method}
\usage{
createAlignmentShellScript(obj, scriptVecSlot = "scriptVec")
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
