% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createAnalysisFolders}
\alias{createAnalysisFolders}
\title{createAnalysisFolders}
\usage{
createAnalysisFolders(obj)
}
\arguments{
\item{obj}{A bioLOGIC data object}
}
\description{
Method to create the analysis folder
}
