% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createAndFormatExcelOutputFiles}
\alias{createAndFormatExcelOutputFiles}
\title{A method}
\usage{
createAndFormatExcelOutputFiles(
  obj,
  metaCoreCountFilter = 1,
  customOutputCols = NULL,
  addedOutputCols = NULL
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
