% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createbulkRNASeqAnalysisBashScripts}
\alias{createbulkRNASeqAnalysisBashScripts}
\title{A method}
\usage{
createbulkRNASeqAnalysisBashScripts(
  obj = "biologic object",
  scriptVecSlot = "scriptVec"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
