% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createbulkRNASeqAnalysisNFcoreScript}
\alias{createbulkRNASeqAnalysisNFcoreScript}
\title{createbulkRNASeqAnalysisNFcoreScript}
\usage{
createbulkRNASeqAnalysisNFcoreScript(
  outdir = "script_output_directory",
  genomeFa = "genomeFa",
  GTFfile = "GTFfile",
  project_id = "RN20202",
  scriptOutdir = NULL,
  NFcoreSettingsFN = "path/to/NF/core/settings",
  scriptVecSlot = "scriptVec"
)
}
\arguments{
\item{outdir}{NF core output directory}

\item{genomeFa}{genomeFa file}

\item{GTFfile}{GTFfile}

\item{project_id}{ProjectID}

\item{scriptOutdir}{Scriptoutputdirectory}

\item{NFcoreSettingsFN}{Path for NF core settings file}

\item{scriptVecSlot}{Script vector slot}
}
\description{
Creates NF-core script
}
