% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createConcatenateFASTQfilesShellScript}
\alias{createConcatenateFASTQfilesShellScript}
\title{A method}
\usage{
createConcatenateFASTQfilesShellScript(
  dfConCat = "dataframe with SRR inut and SRX output file names"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
