% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createDesignFileCrickASFsamples}
\alias{createDesignFileCrickASFsamples}
\title{A method}
\usage{
createDesignFileCrickASFsamples(
  pathToSeqStorageFolder = "pathToSeqStorageFolder",
 
    FNsampleAnnotation = "paste0(\\n    S4LvariableListing$localWorkDir,\\n    \\"sample.ids.txt\\")",
  paired.end = "paired.end",
  fastqDir = "fastqDir",
  baseMount = ""
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
