% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createExcelWorkbook}
\alias{createExcelWorkbook}
\title{createExcelWorkbook}
\usage{
createExcelWorkbook(excelList, outPutFN = "result.table.xlsx")
}
\arguments{
\item{excelList}{A list of dataframes}

\item{outPutFN}{A output filepath/filename}
}
\description{
This function creates an Excel workbook from a list with
data.frames. Each item of the list will become a sheet in the Excel
output file
}
