% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createGeneNameTable}
\alias{createGeneNameTable}
\title{createGeneNameTable}
\usage{
createGeneNameTable(
  obj,
  biomart = "ENSEMBL_MART_ENSEMBL",
  genomeDir = "/camp/svc/reference/Genomics/babs",
  selString = "mmusculus_gene_ensembl",
  host = "http://jan2019.archive.ensembl.org",
  primaryAlignmentGeneID = "ENSMUSG",
  geneIDcolumn = "mgi_symbol"
)
}
\arguments{
\item{obj}{A bioLOGIC data object}
}
\description{
Method to define genome and genename tables
}
