% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createNewProject}
\alias{createNewProject}
\title{createNewProject}
\usage{
createNewProject(
  dbname = "reference_categories_db_new",
  dbtable = "project_description_table",
  password = db.pwd,
  project_name = "G2 Cell Cycle Checkpoint",
  project_lab = ";Parker;",
 
    project_description = "This project investigates the influence of protein kinase C variants on the G2 cell cycle checkpoint."
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
