% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createRNAseqQCscript}
\alias{createRNAseqQCscript}
\title{A method}
\usage{
createRNAseqQCscript(
  obj = "biologic objec",
  scriptVecSlot = "scriptVec",
  bamSuffix = "STAR.genome.bam"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
\keyword{n=n-1',}
