% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createSampleDendrogram}
\alias{createSampleDendrogram}
\title{A method}
\usage{
createSampleDendrogram(obj, Ntop4pca = 500, plotname = "dendrogram")
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
