% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createSettingsFile}
\alias{createSettingsFile}
\title{createSettingsFile}
\usage{
createSettingsFile(
  obj = "object",
  df.data = "database.table",
  publicDataset = FALSE,
  defaultXcolName = NULL,
  defaultYcolName = NULL,
  timepointName = NULL,
  sample.order = "names(database.table)[grep('norm_counts', names(databse.table))]",
  heatmapSampleOrder = "lg2_avg vec",
  sample.names = "",
  count.sample.colors = "rainbow(length(sample.order))",
  ptm.colum = "display_ptm",
  count.table.headline = "PTM ratio H/L counts for all samples",
  count.table.sidelabel = "Counts",
  venn.slider.selector.strings = "c(\\"contrast_x_logFC\\", \\"constrast_x_padj\\")",
 
    plot.selection.strings = "c(\\n    \\"_logFC\\",\\n    \\"_PCA_\\",\\n    \\"_lg10p\\"\\n)",
 
    webSiteDir = "/camp/stp/babs/working/boeings/Stefan/protocol_files/github/biologic/src/experiments",
  upper_heatmap_limit = 3,
  lower_heatmap_limit = -3,
  heamap.headline.text = "heamap.headline.text",
  project_id = "project_id",
  primDataTable = "p123_rna_seq_table",
  pcaDbTable = NULL,
  pointer = "Gene Symbol:"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
