% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createSeuratSettingsFile}
\alias{createSeuratSettingsFile}
\title{createSeuratSettingsFile}
\usage{
createSeuratSettingsFile(
  labname = "labname",
  prim.data.db = "primDataDB",
  cat.ref.db.table = "cat.ref.db.table",
  rnaseqdbTableName = "rnaseqdbTableName",
  geneIDcolumn = "geneIDcolumn",
  displayPTMcolumn = "",
  heatmap.headline.text = "Heatmap",
  count.table.headline = "",
  count.table.sidelabel = "",
  df.data = "database.table",
  defaultXcolName = NULL,
  defaultYcolName = NULL,
  sample.order = "names(database.table)[grep('norm_counts', names(databse.table))]",
  heatmapSampleOrder = "lg2_avg vec",
  sample.names = "",
  count.sample.colors = "rainbow(length(sample.order))",
  ptm.colum = "display_ptm",
  venn.slider.selector.strings = "c(\\"contrast_x_logFC\\", \\"constrast_x_padj\\")",
  plot.selection.strings =
    "c(\\n    \\"_logFC\\",\\n    \\"_PCA_\\",\\n    \\"_lg10p\\"\\n)",
  webSiteDir =
    "/camp/stp/babs/working/boeings/Stefan/protocol_files/github/biologic/src/experiments",
  upper_heatmap_limit = 3,
  lower_heatmap_limit = -3,
  heamap.headline.text = "heamap.headline.text",
  project_id = "project_id",
  primDataTable = "p123_rna_seq_table",
  pcaDbTable = NULL,
  pointer = "Gene Symbol:"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
