% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createSRRdownloadScript}
\alias{createSRRdownloadScript}
\title{createSRRdownloadScript}
\usage{
createSRRdownloadScript(
  srr.id.vector = "srr.id.vector",
  sra.id.vector = "sra.id.vector",
  gse.id.vector = "",
  datadir = "",
 
    module.load.cmd = "module use /camp/stp/babs/working/software/modules/all;module load sratoolkit/2.8.2-1",
  fastqDir = "./",
  project.code = "project.code"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
