% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createTrimGaloreShellScript}
\alias{createTrimGaloreShellScript}
\title{A method}
\usage{
createTrimGaloreShellScript(obj, scriptVecSlot = "scriptVec")
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
