% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{datatable.to.website.ptm}
\alias{datatable.to.website.ptm}
\title{A method}
\usage{
datatable.to.website.ptm(
  df.data,
  gene.id.column = "ENSMUSG",
  heatmap.genes = "",
  n.cluster.genes = 6000,
  count.data = FALSE,
  logFC.cut.off = 0,
  selector4heatmap.cols = "logFC",
  heatmap.preprocessing = "lg2.row.avg",
  hm.cut.off = 4,
  n.hm.cluster = 10,
  count.cut.off.filter = 1
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
