% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{DGEanalysis}
\alias{DGEanalysis}
\title{DGEanalysis}
\usage{
DGEanalysis(obj, createNewResultTable = TRUE)
}
\arguments{
\item{obj}{A biologic object}
}
\description{
Method description
}
\keyword{TBD}
