% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{do.differential.expression.analyis}
\alias{do.differential.expression.analyis}
\title{A method}
\usage{
do.differential.expression.analyis(
  raw.counts.filt = raw.counts.filt,
  DEseq2Dir = paste0(localWorkDir, "DESeq2"),
  df.design = df.design,
  gene.id = "ENSMUSG",
  batch.mode = FALSE,
  parallel = FALSE,
  toBeNormalized = TRUE,
  plotOutputDir = "",
  doPCA = TRUE,
  writePlotsToFile = TRUE,
  timeseries = FALSE,
  tempShellScriptVector = as.vector(NULL, mode = "character")
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
