% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{doDGEanalysis}
\alias{doDGEanalysis}
\title{A method}
\usage{
doDGEanalysis(
  obj,
  DGEdesignCols = "names(Obio@dfDesign)[grep(\\"comp_\\", names(Obio@dfDesign)) ]",
  createNewResultTable = TRUE,
  normaliseAllSamplesTogether = FALSE
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
