% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{doLinearFittings}
\alias{doLinearFittings}
\title{A method}
\usage{
doLinearFittings(
  obj,
  designColSelector = "",
  mode = "independentVariation",
  Ntop4pca = 500,
  plotname = "plotname"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
