% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{doLRTanalysis}
\alias{doLRTanalysis}
\title{A method}
\usage{
doLRTanalysis(
  obj,
  LRTdesignCols = "names(Obio@dfDesign)[grep(\\"LRT_\\", names(Obio@dfDesign)) ]",
  createNewResultTable = TRUE
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
