% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{doPCA}
\alias{doPCA}
\title{A method}
\usage{
doPCA(
  obj,
  Ntop4pca = 500,
  avgCountCutOffperSample = 0,
  pcaSelectionVec = NULL,
  pcaDimensionsToInvestigate = c(1:5)
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
