% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{doQuery}
\alias{doQuery}
\title{doQuery}
\usage{
doQuery(
  user = "db.user",
  password = "db.upload.pwd",
  host = "host",
  dbname = "primDataDB",
  query = "mysql db query",
  resOut = FALSE,
  geneDefault = NULL,
  mode = "MySQL"
)
}
\arguments{
\item{user}{}

\item{password}{}

\item{This}{can be MySQL or SQLite}
}
\value{
Res
}
\description{
doQuery
}
