% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{hpcClusterSubmisison}
\alias{hpcClusterSubmisison}
\title{A method}
\usage{
hpcClusterSubmisison(
  CMDstring = "ls",
  jobname = "SBhpcRun",
  cores = 1,
  memPerCpu = 7000,
  logFN = "report.file.slurm",
  commands.text.file = "commands.txt"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
