% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{import.db.table.from.db}
\alias{import.db.table.from.db}
\title{A method}
\usage{
import.db.table.from.db(
  dbtable = "interpro_categori",
  dbname = "reference_categories_db_new",
  user = "boeings",
  password = "",
  host = "www.biologic-db.org",
  port = 6008
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
