% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{list.db.table.col.names}
\alias{list.db.table.col.names}
\title{list.db.table.col.names}
\usage{
list.db.table.col.names(
  dbtable = "interpro_categori",
  dbname = "reference_categories_db_new",
  user = "boeings",
  password = "",
  host = "www.biologic-db.org",
  port = 6008
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
