% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{list.db.tables.in.db}
\alias{list.db.tables.in.db}
\title{list.db.tables.in.db}
\usage{
list.db.tables.in.db(
  dbname = "reference_categories_db_new",
  user = "boeings",
  password = "",
  host = "www.biologic-db.org"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
