% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{listExistingProjects}
\alias{listExistingProjects}
\title{listExistingProject}
\usage{
listExistingProjects(
  user = "boeings",
  password = "",
  host = "www.biologic-db.org",
  port = 6008,
  dbname = "reference_categories_db_new",
  dbtable = "project_description_table"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
