% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{loadDESeq2outputFromFile}
\alias{loadDESeq2outputFromFile}
\title{loadDESeq2outputFromFile}
\usage{
loadDESeq2outputFromFile(
  Obio,
  DEseq2resultDir = "Obio@parameterList$DEseq2External_DGE",
  replace = FALSE,
  mode = NULL
)
}
\arguments{
\item{Obio}{biologic object}

\item{DEseq2resultDir}{Path to DEseq2 results(dds) output directory. Make sure the last character is a /}

\item{mode}{LRT or DGE}
}
\value{
Obio
}
\description{
This function creates a settings file for a biologic data visualization project.
}
