% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{make.hm}
\alias{make.hm}
\title{A method}
\usage{
make.hm(
  m.df1,
  filename = "heatmap",
  k.number = 10,
  n.colors = 1000,
  hclust.method = "complete",
  dist.method = "euclidean",
  main = "",
  Colv = TRUE,
  showRowNames = FALSE,
  showColNames = TRUE,
  plotSeparationLines = FALSE
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
