% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{msigdb.gmt2refDB}
\alias{msigdb.gmt2refDB}
\title{msigdb.gmt2refDB}
\usage{
msigdb.gmt2refDB(
  df.gmt = "read.delim(gmt.file, header=FALSE, sep = '\\t', stringsAsFactors = FALSE)",
  host = "www.biologic-db.org",
  db.user = "boeings",
  pwd = "pwd",
  ref.db = "reference_database",
  ref.db.table = "reference database table name",
  cat_type = "mysigdb_c5",
  data_source = "Broad Institute",
  keep.gene.values = FALSE,
  gene.id = "hgnc_symbol",
  create.new.table = FALSE,
  mm.hs.conversion.file = "C:/Users/boeing01/Desktop/homologene.data"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
