% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{organizeFastqFiles}
\alias{organizeFastqFiles}
\title{A method}
\usage{
organizeFastqFiles(
  baseMount = "gsub(\\"boeings/\\", \\"\\", hpc.mount)",
  pathToSeqStorageFolder = "pathToSeqStorageFolder",
  fastqOutputDir = "fastqDir",
  localWorkDir = "localworkdir",
  checkFileExists = TRUE
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
