% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{prepareRSEMcountMatrix}
\alias{prepareRSEMcountMatrix}
\title{A method}
\usage{
prepareRSEMcountMatrix(
  obj = "biologic objec",
  string.to.be.deleted.in.raw.counts.columns = paste0("X", gsub("/", ".",
    paste0(obj@parameterList$workdir, "RSEM/Ensembl/")))
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
