% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{profileCluster}
\alias{profileCluster}
\title{profileCluster}
\usage{
profileCluster(
  obj = "Obio",
  markerList = "residualClusterMarkers",
  gmtList = "clusterProfilerGMTList",
  nTop = 10,
  pvalueCutoff = 0.5
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
