% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{readAndPrepareCountMatrix}
\alias{readAndPrepareCountMatrix}
\title{A method}
\usage{
readAndPrepareCountMatrix(
  count.data.fn = "paste0(raw.count.dir, \\"/\\",count.data.file)",
  string.to.be.deleted.in.raw.counts.columns =
    "paste0(\\"X\\", gsub(\\"/\\", \\".\\",paste0(workdir, \\"RSEM/Ensembl/\\")))",
  df.design = "dfDesign"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
