% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{retrieve.gene.category.from.db}
\alias{retrieve.gene.category.from.db}
\title{retrieve.gene.category.from.db}
\usage{
retrieve.gene.category.from.db(
  cat_id = "mysigdb_c5_MF__127",
  dbname = "reference_categories_db_new",
  user = "boeings",
  password = "",
  host = "www.biologic-db.org",
  gene.symbol = "mgi_symbol",
  print.cat.name = TRUE
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
