% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{samplesToExcludeFromTSdisplay}
\alias{samplesToExcludeFromTSdisplay}
\title{sampleToExcludeFromTSdisplay}
\usage{
samplesToExcludeFromTSdisplay(Obio, designTScol = "timepointName")
}
\arguments{
\item{Obio}{An Obio object}

\item{designTScol}{timeseries column}
}
\description{
This function identifies samples that need to be dropped from a timecoures display
}
