% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{selectHeatmapGenes}
\alias{selectHeatmapGenes}
\title{A method}
\usage{
selectHeatmapGenes(
  dfData = "df.summary",
  selCol = "logFC",
  cutOff = 0.75,
  pSelCol = "padj",
  pCutOff = 1,
  zeroOneCol = "logFC_cut_off",
  geneID = "hgnc_symbol"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
