% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{setCrickGenomeAndGeneNameTable}
\alias{setCrickGenomeAndGeneNameTable}
\title{setCrickGenomeAndGeneNameTable}
\usage{
setCrickGenomeAndGeneNameTable(
  obj,
  genomeDir = "/camp/svc/reference/Genomics/babs"
)
}
\arguments{
\item{obj}{A bioLOGIC data object}
}
\description{
Method to define genome and genename tables
}
