% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{setDataBaseParameters}
\alias{setDataBaseParameters}
\title{setDataBaseParameters}
\usage{
setDataBaseParameters(obj)
}
\arguments{
\item{obj}{A bioLOGIC data object}
}
\description{
Method to set base parameters
}
