% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{setGenomeAndGeneNameTable}
\alias{setGenomeAndGeneNameTable}
\title{setGenomeAndGeneNameTable}
\usage{
setGenomeAndGeneNameTable(
  obj,
  genome = "GRCm38",
  primaryAlignmentGeneID = "ENSMUSG",
  release = NULL,
  path2GeneIDtable =
    "paste0(obj@parameterList$hpcMount, 'Projects/reference_data/gene_id_annotation_files/','20171206.release-89.mm.ENSMUSG.mgi.entrez.uniprot.description.hgnc.table.txt')",
  GTFfile =
    "paste0('/camp/svc/reference/Genomics/babs/mus_musculus/ensembl/GRCm38/',obj@parameterList$release,'/gtf/Mus_musculus.GRCm38.', releaseID, '.rnaseqc.gtf')",
  genomeFa =
    "paste0('/camp/svc/reference/Genomics/babs/mus_musculus/ensembl/GRCm38/',obj@parameterList$release,'/genome/Mus_musculus.GRCm38.dna_sm.toplevel.fa')",
  genomeFai =
    "paste0('/camp/svc/reference/Genomics/babs/mus_musculus/ensembl/GRCm38/',obj@parameterList$release,'/genome/Mus_musculus.GRCm38.dna_sm.toplevel.fai')",
  rRNAfile =
    "paste0('/camp/svc/reference/Genomics/babs/mus_musculus/ensembl/GRCm38/', obj@parameterList$release, '/gtf/Mus_musculus.GRCm38.',releaseID,'.rRNA.list')",
  geneIDcolumn = "mgi_symbol",
  bedFile =
    "paste0('/camp/svc/reference/Genomics/babs/mus_musculus/ensembl/GRCm38/', obj@parameterList$release,'/gtf/Mus_musculus.GRCm38.',releaseID,'.bed')",
  refFlatFile = NULL,
  ribosomalIntervalList = NULL,
  genomeidx = NULL,
  bowtieGenomeidx = NULL
)
}
\arguments{
\item{obj}{A bioLOGIC data object}

\item{genome}{Genome ID, e.g. GRCm38}
}
\description{
Method to define genome and genename tables
}
