% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{setMountingPoint}
\alias{setMountingPoint}
\title{setMountingPoint}
\usage{
setMountingPoint(obj)
}
\arguments{
\item{obj}{A bioLOGIC object}
}
\description{
Method to set mounting points. Setting the moutning point according to the local environment.
}
