% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{upload.pca.table.to.db}
\alias{upload.pca.table.to.db}
\title{A method}
\usage{
upload.pca.table.to.db(
  df.pca,
  host = "www.biologic-db.org",
  prim.data.db = "vtl_data",
  password = "",
  db.user = "boeings",
  PCAdbTableName = "P79_VTL_ES_PCA"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
