% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{addGeneAnnotation}
\alias{addGeneAnnotation}
\title{addGeneAnnotation}
\usage{
addGeneAnnotation(obj, geneIDfn = NULL, addUniprotColumn = FALSE)
}
\arguments{
\item{obj}{A bioLOGIC data object}

\item{geneIDfn}{GeneID definiton file. Expected columns: primaryAlignmentGeneID, geneIDcolumn optional addition: hgnc_symbol, gene_description, gene_type}

\item{addUniprotColumn}{Boolean: Adding a uniprot column?}
}
\description{
Method to add a gene annotation
}
