% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{assembleBiologicProject}
\alias{assembleBiologicProject}
\title{assembleBiologicProject}
\usage{
assembleBiologicProject(
  designFN = NULL,
  modelFN = NULL,
  NFcoreSettingsFN = NULL,
  countTableFN = NULL,
  TpmTableFN = NULL,
  PcaFN = NULL,
  mostVariableFeaturesFN = NULL,
  calculate_DGE = TRUE,
  calculate_LRT = TRUE,
  DEseq2External_DGE = NULL,
  DEseq2External_LRT = NULL,
  projectFolder = NULL,
  experiment_id = NULL,
  title = NULL,
  subtitle = NULL,
  abstract = NULL,
  project_name = NULL,
  lims.id = NULL,
  labname = NULL,
  NtopGenes = NULL,
  corGeneVec = NULL,
  experiment.type = NULL,
  species = NULL,
  release = "release-89",
  count.table.headline = "TPM-values for all Samples",
  count.table.sidelabel = "TPM",
  heatmap.headline.text = "Heatmap: Row-averaged Expr",
  designTScol = NULL,
  timecourse.units = NULL,
  primDataDB = NULL,
  db.user = NULL,
  host = NULL,
  lab.categories.table = NULL,
  HGtestEnrichmentGmtFile = NULL,
  GSEAtestEnrichmentGmtFile = NULL,
  pathToSeqStorageFolder = NULL,
  stranded = FALSE,
  read.length = "100bp",
  paired.end = FALSE,
  pathToRSEMresultsFiles = NULL,
  dfRef = NULL
)
}
\arguments{
\item{designFN}{Path to a design file.}

\item{modelFN}{Path to DGE model file}

\item{NFcoreSettingsFN}{Path to NF core settings file.}

\item{countTableFN}{Path to count table}

\item{TpmTableFN}{Path to TPM table}

\item{PcaFN}{Path to PCA table}

\item{mostVariableFeaturesFN}{Path to most variable featurs file}

\item{calculate_DGE}{If DGE instructions are present in the desing file, calculate}

\item{calculate_LRT}{If LRT instructions are present in the desing file, calculate}

\item{DEseq2External_DGE}{Folder in which external res(dds) results are kept}

\item{DEseq2External_LRT}{Folder in which external res(dds) results are kept}

\item{projectFolder}{Path to project folder}

\item{experiment_id}{Experiment ID}

\item{title}{Experiment title}

\item{subtitle}{Experiment subtitle - usually labnane - lead scientis}

\item{abstract}{Abstract}

\item{project_name}{Project name}

\item{lims.id}{LIMS ID to retrieve meta data}

\item{labname}{Labname}

\item{NtopGenes}{Ntopgenes to use for PCA}

\item{corGeneVec}{Vector with genes to calculate correlations for}

\item{experiment.type}{Experiment type:  options: "bulk_rna_seq", "sc_rna_seq", "mi_rna_seq"}

\item{species}{Species. Options: "homo_sapiens", "mus_musculus", "drosophila_melongaster", "daniero_rerio"}

\item{release}{Ensembl release used for annotations. Option: "release-89",}

\item{count.table.headline}{String to be displayed in count table headlines}

\item{count.table.sidelabel}{String to be displayed as count table sidelabel}

\item{heatmap.headline.text}{Heatmap headline text}

\item{designTScol}{Column name of a numeric column in the design file}

\item{timecourse.units}{Timecourse units}

\item{primDataDB}{Primary database}

\item{db.user}{Database username}

\item{host}{Database host IP}

\item{lab.categories.table}{Lab categories table}

\item{HGtestEnrichmentGmtFile}{HGtestEnrichmentGmtFile}

\item{GSEAtestEnrichmentGmtFile}{GSEAtestEnrichmentGmtFile}

\item{pathToSeqStorageFolder}{Vector of paths to folders to Crick sequencing storage. Make sure the last character is a /}

\item{stranded}{Boolean}

\item{read.length}{default 100bp}

\item{paired.end}{Boolean}

\item{pathToRSEMresultsFiles}{pathToRSEMresultsFiles}

\item{dfRef}{Dataframe for reference files}

\item{excelList}{A list of dataframes}

\item{outPutFN}{A output filepath/filename}
}
\description{
This function creates a settings file for a biologic data visualization project.
}
