% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{setCrickGenomeAndGeneNameTable}
\alias{setCrickGenomeAndGeneNameTable}
\title{A method}
\usage{
setCrickGenomeAndGeneNameTable(
  obj,
  genomeDir = "/camp/svc/reference/Genomics/babs"
)
}
\arguments{
\item{agree}{TBD}
}
\description{
Method description
}
\keyword{TBD}
